Fix serialization dependencies
