+++ Packages/org.nahoum.easy-web-interop/Runtime/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85913a25205196f41a30632a6f4c49e2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
