+++ Packages/org.nahoum.easy-web-interop/Runtime/Serialization/BasicCollectionSerializer.cs
