+++ Packages/org.nahoum.easy-web-interop/Runtime/Serialization/ExposeWebSerializerAttribute.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7874219156c217548842bbc334dd8311
